Many new and improved move generators and evaluators.

+++ src/codecup2018/evaluator/Evaluator.java
@@ -12,5 +12,5 @@ public interface Evaluator {
      * @param board
      * @return
      */
-    public abstract double evaluate(Board board);
+    public abstract int evaluate(Board board);
 }
